+++ memory/__init__.py
@@ -20,7 +20,7 @@ from memory.embeddings.autoencoder import (
 )
 # from memory.embeddings.text_embeddings import TextEmbeddingEngine
 from memory.embeddings.compression import CompressionEngine
-from memory.memory_agent import MemoryAgent
+from memory.agent_memory import MemoryAgent
 
 # Retrieval components
 from memory.retrieval.attribute import AttributeRetrieval
