+++ tests/debug_memory_transitions.py
@@ -7,7 +7,7 @@ import time
 from typing import Dict, Any
 
 from memory.config import MemoryConfig
-from memory.memory_agent import MemoryAgent
+from memory.agent_memory import MemoryAgent
 
 # Setup logging
 logging.basicConfig(level=logging.DEBUG)
